templates/others/flask-app-4.1-csrf/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
pymongo==4.2.0
pymongo[srv]
mongoengine==0.24.1
Flask-JWT-Extended==4.4.4
Werkzeug==2.3.4 
Flask-Limiter
Flask-Mail
boto3
python-dotenv